Fixes to compare_dict to more fully do the comparison
I found that how compare_dict was being called may have changed as
it was not fully comparing the dictionary elements.  I've updated
this version to do a more complete key/value comparison in both
directions as well as summarizing all of the failures rather than
alerting at just the first one encountered.

Ref ntc-templates/#73